+++ src/Admin/ErrorLogAdmin.php
@@ -12,13 +12,6 @@ use Sonata\AdminBundle\Show\ShowMapper;
 
 class ErrorLogAdmin extends AbstractAdmin
 {
-    /** @var array */
-    protected $datagridValues = [
-        '_page' => 1,
-        '_sort_order' => 'DESC',
-        '_sort_by' => 'date_created',
-    ];
-
     public function configureListFields(ListMapper $list)
     {
         $list
@@ -37,6 +30,13 @@ class ErrorLogAdmin extends AbstractAdmin
                     ],
                 ]
             );
+    }
+
+    protected function configureDefaultSortValues(array &$sortValues): void
+    {
+        parent::configureDefaultSortValues($sortValues);
+        $sortValues[DatagridInterface::SORT_BY] = 'date_created';
+        $sortValues[DatagridInterface::SORT_ORDER] = 'DESC';
     }
 
     protected function configureShowFields(ShowMapper $show)